refactor: share
