--- src/common/interfaces/dto/share/share.request.dto.ts
@@ -1,20 +1,0 @@
-import { Type } from 'class-transformer';
-import { IsInt, IsNumber, IsOptional, IsString } from 'class-validator';
-
-export class TimetableImageQueryDto {
-  @Type(() => Number)
-  @IsNumber()
-  timetable!: number;
-
-  @Type(() => Number)
-  @IsNumber()
-  year!: number;
-
-  @Type(() => Number)
-  @IsNumber()
-  semester!: number;
-
-  @IsString()
-  @IsOptional()
-  language?: string;
-}
